+++ pyproject.toml
@@ -13,6 +13,7 @@ psycopg2 = "^2.9.9"
 boto3 = "^1.34.59"
 python-dotenv = "^1.0.1"
 python-multipart = "^0.0.9"
+awscli = "^1.32.59"
 
 
 [build-system]
